Separate file for local flexible
